% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_screen}
\alias{add_screen}
\title{Adds a new screen to a list of screens}
\usage{
add_screen(
  screen_list = NULL,
  name,
  replicates,
  normalize_name = NULL,
  target_coverage = 200
)
}
\arguments{
\item{screen_list}{An existing list of screens (optional, default NULL).}

\item{name}{A name for the screen (e.g. "RPE1_T18").}

\item{replicates}{A list of columns containing replicate data for the given screen.}

\item{normalize_name}{A name of another screen to normalize data in this screen to
(e.g. "RPE1_T0", optional, default NULL).}

\item{target_coverage}{Target coverage of the given screen (optional, default 100).}
}
\value{
A named list corresponding to provided screen names, where each sub-list 
  contains a list of the replicate columns (in "replicates") and the screen to 
  normalize against (in "normalize_name").
}
\description{
Makes a list containing info for a given screen and optionally appends it to a given list.
}
