% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_screens_from_table}
\alias{add_screens_from_table}
\title{Adds a list of screens from a sample table file}
\usage{
add_screens_from_table(table_file)
}
\arguments{
\item{table_file}{A tab-separated sample table with three columns, described above: 
Screen, Replicates, and NormalizeTo. Screen is the name of the screen, replicates 
are each technical replicate separated by semicolons, and NormalizeTo is either 
the screen to normalize against or NA if unnecessary (e.g. for T0 screens).}
}
\value{
A named list corresponding to provided screen names, where each sub-list 
  contains a list of the replicate columns (in "replicates") and the screen to 
  normalize against (in "normalize_name").
}
\description{
For a sample table formatted as a .tsv file with columns for screen, replicates,
and the name of a screen to compute log fold-changes against, adds each screen
to a list and returns that list.
}
