% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{call_combn_hits}
\alias{call_combn_hits}
\title{Call significant responses for scored paired data.}
\usage{
call_combn_hits(
  scores,
  screen_names,
  fdr_threshold = 0.1,
  differential_threshold = 0.5,
  neg_type = "Negative",
  pos_type = "Positive"
)
}
\arguments{
\item{scores}{Dataframe returned from \code{score_combn_vs_single}.}

\item{screen_names}{A list of screen names scored in \code{score_combn_vs_single}
for which to call significant effects.}

\item{fdr_threshold}{Threshold below which to call gene effects as significant 
(default 0.1).}

\item{differential_threshold}{Absolute value threshold on differential effects, 
below which gene effects are not called as significant (default 0.5).}

\item{neg_type}{Label for significant effects with a negative differential effect
(default "Negative").}

\item{pos_type}{Label for significant effects with a positive differential effect
(default "Positive").}
}
\value{
Dataframe of scored data with differential effects called as significant
  for the specified conditions.
}
\description{
Run this to call significant responses for data returned from 
\code{score_combn_vs_single}. Do NOT run this on data returned from
\code{score_conditions_vs_control}.
}
