% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{call_condition_hits}
\alias{call_condition_hits}
\title{Call significant responses for scored data.}
\usage{
call_condition_hits(
  scores,
  control_screen_name,
  condition_screen_names,
  fdr_threshold = 0.1,
  differential_threshold = 0,
  neg_type = "Negative",
  pos_type = "Positive",
  fdr_method = "BH",
  expected_guides = 6
)
}
\arguments{
\item{scores}{Dataframe returned from \code{score_conditions_vs_control}.}

\item{control_screen_name}{Name of a control screen to test condition screens against.}

\item{condition_screen_names}{A list of condition screen names to score against the 
control screen.}

\item{fdr_threshold}{Threshold below which to call gene effects as significant 
(default 0.1).}

\item{differential_threshold}{Absolute value threshold on differential effects, 
below which gene effects are not called as significant (default 0).}

\item{neg_type}{Label for significant effects with a negative differential effect
(default "Negative").}

\item{pos_type}{Label for significant effects with a positive differential effect
(default "Positive").}

\item{fdr_method}{Type of FDR to compute. One of "BH", "BY" or "bonferroni" (default
"BY")}

\item{expected_guides}{Number of guides expected per gene-pair, to more accurately
assign a hit quality metric (default 15).}
}
\value{
Dataframe of scored data with differential effects called as significant
  for the specified conditions.
}
\description{
Run this to call significant responses for data returned from 
\code{score_conditions_vs_control}. Do NOT run this on data returned from
\code{score_combn_vs_single}.
}
