% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{call_significant_response}
\alias{call_significant_response}
\title{Call significant responses for scored data.}
\usage{
call_significant_response(
  scores,
  control_cols,
  ...,
  fdr_threshold = 0.1,
  differential_threshold = 0,
  neg_type = "Negative",
  pos_type = "Positive"
)
}
\arguments{
\item{scores}{Dataframe returned from \code{score_conditions_vs_control}.}

\item{control_cols}{A list where the first entry is a desired name for a condition, 
e.g. DMSO, and all subsequent values are column names of the technical replicates 
for that condition. See documentation for \code{score_conditions_vs_control}.}

\item{...}{Any number of conditions to score against the specified control, passed 
in as separate lists formatted the same as control_cols.}

\item{fdr_threshold}{Threshold below which to call gene effects as significant 
(default 0.1).}

\item{differential_threshold}{Absolute value threshold on differential effects, 
below which gene effects are not called as significant (default 0).}

\item{neg_type}{Label for significant effects with a negative differential effect
(default "Negative").}

\item{pos_type}{Label for significant effects with a positive differential effect
(default "Positive").}
}
\value{
Dataframe of scored data with differential effects called as significant
  for the specified conditions.
}
\description{
Run this to call significant responses for data returned from 
\code{score_conditions_vs_control}. Do NOT run this on data returned from
\code{score_dual_vs_single}.
}
