% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_batch_file}
\alias{check_batch_file}
\title{Checks a batch scoring file}
\usage{
check_batch_file(batch_file, screens)
}
\arguments{
\item{batch_file}{Path to .tsv file mapping screens to their controls for scoring, with two 
columns for "Screen" and "Control." Screens to score against derived null-models with the
combn scoring mode must have their respective control labeled as "combn."}
}
\value{
TRUE.
}
\description{
Checks to make sure that the contents of the .tsv file and its formats are appropriate for
running batch scoring functions.
}
\keyword{internal}
