% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_screen_params}
\alias{check_screen_params}
\title{Checks input parameters}
\usage{
check_screen_params(df, screens)
}
\arguments{
\item{df}{Reads or LFC dataframe.}

\item{screens}{List of screens created with \code{add_screens}.}
}
\value{
TRUE.
}
\description{
Checks to make sure that the given screens match the given dataframe.
}
\keyword{internal}
