% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chymera_batch_table}
\alias{chymera_batch_table}
\title{Batch table for chymera_paralog}
\format{
A dataframe formatted as described in \code{\link{score_combn_batch}}.
\describe{
 \item{Screen}{name of screen to score}
 \item{Control}{name of screen to compare against, or "combn" to score combinatorial guides}
}
}
\usage{
chymera_batch_table
}
\description{
Batch table for chymera_paralog
}
\examples{
screens <- add_screens_from_table(chymera_sample_table)
load_batch_table(chymera_batch_table, screens)
}
\keyword{datasets}
