% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chymera_sample_table}
\alias{chymera_sample_table}
\title{Sample table for chymera_paralog}
\format{
A dataframe formatted as described in \code{\link{add_screens_from_table}}.
\describe{
 \item{Screen}{name of screens to score}
 \item{Replicates}{associated technical replicates for each screens as columns of 
                   chymera_paralog}
 \item{NormalizeTo}{name of screen to normalize against}
}
}
\usage{
chymera_sample_table
}
\description{
Sample table for chymera_paralog
}
\examples{
add_screens_from_table(chymera_sample_table)
}
\keyword{datasets}
