% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{essential_lfc_qc}
\alias{essential_lfc_qc}
\title{Computes essential gene recovery AUC.}
\usage{
essential_lfc_qc(df, screens, gene_col1, gene_col2, output_folder)
}
\arguments{
\item{df}{LFC dataframe.}

\item{screens}{List of screens generated with \code{add_screens}.}

\item{gene_col1}{A column containing the first of two gene names. Intergenic regions must be 
denoted as "---" and non-targeted genes must be denoted as "None".}

\item{gene_col2}{See above.}

\item{output_folder}{Folder to which essential gene QC results should be saved.}
}
\value{
All output is saved to the file "essential_PR_QC.txt" in the folder given by
  output_folder.
}
\description{
Computes area under the curve for ROC curves that measure how well each technical replicate
recovers signal for essential-targeting guides and saves results to file. Only computes AUC
for guides that target essential genes twice, guides that target two different essential 
genes, or guides that target an essential gene and an intergenic region.
}
