% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{filter_reads}
\alias{filter_reads}
\title{Filters guides with too few read counts.}
\usage{
filter_reads(df, cols, min_reads = 30, max_reads = 10000)
}
\arguments{
\item{df}{Reads dataframe.}

\item{cols}{Columns to filter by.}

\item{min_reads}{Minimum number of reads to keep (anything below
this value will be filtered out).}

\item{max_reads}{Maximum number of reads to keep (anything above
this value will be filtered out).}
}
\value{
Filtered dataframe.
}
\description{
Filters guides out with too few read counts from a given reads dataframe
and a given set of columns (e.g. T0 columns).
}
