% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_batch_table}
\alias{load_batch_table}
\title{Checks a batch scoring file}
\usage{
load_batch_table(batch_table, screens)
}
\arguments{
\item{batch_table}{Either a dataframe or a path to .tsv file mapping screens to their controls 
for scoring, with two columns for "Screen" and "Control." Screens to score against derived 
null-models with the combn scoring mode must have their respective control labeled as "combn."}

\item{screens}{A screens object created with \code{add_screens_from_table}.}
}
\value{
A dataframe with Screen and Control columns.
}
\description{
Checks to make sure that the contents of the .tsv file and its formats are appropriate for
running batch scoring functions. The batch table can be either a path to a .tsv file or a 
dataframe with the columns mentioned below.
}
