% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{normalize_screens}
\alias{normalize_screens}
\title{Normalizes reads for given screens}
\usage{
normalize_screens(
  df,
  screens,
  filter_names = NULL,
  scaling_factor = 1e+06,
  pseudocount = 1,
  min_reads = 30,
  max_reads = 10000
)
}
\arguments{
\item{df}{Reads dataframe.}

\item{screens}{List of screens generated with \code{add_screens}.}

\item{filter_names}{List of screen names to filter based on read counts by.}

\item{scaling_factor}{Scaling factor for normalizing reads, helps make different
replicates comparable but the specific choice typically has little to no 
impact on results (default 1e6).}

\item{pseudocount}{Pseudocount for normalized reads, where lower values are
such as 1-10 are recommended for most combinatorial screens with sufficient
coverage, e.g. 200x and upwards (default 1).}

\item{min_reads}{Minimum number of reads to keep (default 30, anything
below this value will be filtered out).}

\item{max_reads}{Maximum number of reads to keep (default 10000, anything
above this value will be filtered out).}
}
\value{
Normalized dataframe.
}
\description{
Log2 and depth-normalizes reads between a given list of columns and a given
column of an earlier timepoint (e.g. T0) specified in the "normalize_name"
entry of each screen. Screens with NULL for their "normalize_name" entry
are log2 and depth-normalized, but not normalized to earlier timepoints.
If a screen to normalize against has multiple replicates, those replicates
are averaged before normalization. Multiple replicates for a screen being
normalized, however, are normalized separately against the provided early
timepoint.
}
