% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_combn_gene_residuals}
\alias{plot_combn_gene_residuals}
\title{Plot guide-level residuals for a given gene.}
\usage{
plot_combn_gene_residuals(scores, residuals, gene1, gene2, condition_name)
}
\arguments{
\item{scores}{Dataframe of scores returned from \code{call_combn_hits}.}

\item{residuals}{Residuals returned with the return_residuals argument set to true
from \code{call_combn_hits}.}

\item{gene1}{First gene name for guides to plot.}

\item{gene2}{Second gene name for guides to plot.}

\item{condition_name}{Name of condition passed to \code{call_combn_hits}.}
}
\value{
A ggplot object.
}
\description{
Plots guide-level residuals for a given gene pair and returns the plot. Works for data 
returned from \code{call_combn_hits}.
}
