% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_combn_lfc}
\alias{plot_combn_lfc}
\title{Plot LFCs for all gene pairs.}
\usage{
plot_combn_lfc(
  scores,
  residuals,
  condition_name,
  output_folder,
  neg_type = "Negative",
  pos_type = "Positive",
  plot_type = "png"
)
}
\arguments{
\item{scores}{Dataframe of scores returned from \code{call_combn_hits}.}

\item{residuals}{Residuals returned with the return_residuals argument set to true
from \code{call_combn_hits}.}

\item{condition_name}{Name of condition passed to \code{call_combn_hits}.}

\item{output_folder}{Folder to output plots to.}

\item{neg_type}{Label for significant effects with a negative differential effect
passed to \code{call_combn_hits} (default "Negative").}

\item{pos_type}{Label for significant effects with a positive differential effect
passed to \code{call_combn_hits} (default "Positive").}

\item{plot_type}{Type of plot to output, one of "png" or "pdf" (default "png").}
}
\description{
Plots replicate comparisons for all replicates in a list of screens and outputs
plots to a given folder. Works for data returned from \code{call_combn_hits}.
}
