% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_combn_response}
\alias{plot_combn_response}
\title{Plots drug response for scored paired data.}
\usage{
plot_combn_response(
  scores,
  condition_name,
  filter_names = NULL,
  loess = TRUE,
  neg_type = "Negative",
  pos_type = "Positive"
)
}
\arguments{
\item{scores}{Dataframe of scores returned from \code{call_combn_hits}.}

\item{condition_name}{Name of condition passed to \code{call_combn_hits}.}

\item{filter_names}{If a list of column names is given, calls points as non-significant 
if they are significant in the provided columsn (e.g. to remove points significant 
in control screens; default NULL).}

\item{loess}{If true and data was loess-normalized, plots loess null model instead
(default TRUE).}

\item{neg_type}{Label for significant effects with a negative differential effect
passed to \code{call_combn_hits} (default "Negative").}

\item{pos_type}{Label for significant effects with a positive differential effect
passed to \code{call_combn_hits} (default "Positive").}
}
\value{
A ggplot object.
}
\description{
Pretty-plots response for data which uses a derived null model (e.g. for comparing
dual-gene knockout effects to a multiplicative null model derived from single-gene
effects). Assumes that data was scored by \code{score_combn_vs_single} and 
significant effects were called by \code{call_combn_hits}.
}
