% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_condition_residuals}
\alias{plot_condition_residuals}
\title{Plot guide-level residuals for all gene pairs.}
\usage{
plot_condition_residuals(
  scores,
  residuals,
  control_name,
  condition_name,
  output_folder,
  neg_type = "Negative",
  pos_type = "Positive",
  plot_type = "png"
)
}
\arguments{
\item{scores}{Dataframe of scores returned from \code{call_condition_hits}.}

\item{residuals}{Residuals returned with the return_residuals argument set to true
from \code{call_condition_hits}.}

\item{control_name}{Name of control passed to \code{call_condition_hits}.}

\item{condition_name}{Name of condition passed to \code{call_condition_hits}.}

\item{output_folder}{Folder to output plots to.}

\item{neg_type}{Label for significant effects with a negative differential effect
passed to \code{call_condition_hits} (default "Negative").}

\item{pos_type}{Label for significant effects with a positive differential effect
passed to \code{call_condition_hits} (default "Positive").}

\item{plot_type}{Type of plot to output, one of "png" or "pdf" (default "png").}
}
\description{
Plots replicate comparisons for all replicates in a list of screens and outputs
plots to a given folder. Works for data returned from \code{call_condition_hits}.
}
