% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plots Pearson correlation heatmap.}
\usage{
plot_heatmap(df, col_groups, filename, display_numbers)
}
\arguments{
\item{df}{Reads or lfc dataframe}

\item{col_groups}{List of grouping labels for each column in df.}

\item{filename}{Output filename for plot.}

\item{display_numbers}{Whether or not to include PCC values in heatmap (default TRUE).}
}
\value{
Writes plot to file with no return value.
}
\description{
Plots heatmap of Pearson correlations for a dataframe with labels
for groups of samples and a blue-yellow color scheme.
}
