% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_lfc}
\alias{plot_lfc}
\title{Plot LFCs for all gene pairs.}
\usage{
plot_lfc(
  scores,
  residuals,
  control_name,
  condition_name,
  output_folder,
  neg_type = "Negative",
  pos_type = "Positive"
)
}
\arguments{
\item{scores}{Dataframe of scores returned from \code{call_significant_response}.}

\item{residuals}{Residuals returned with the return_residuals argument set to true
from \code{call_significant_response}.}

\item{control_name}{Name of control passed to \code{call_significant_response}.}

\item{condition_name}{Name of condition passed to \code{call_significant_response}.}

\item{output_folder}{Folder to output plots to.}

\item{neg_type}{Label for significant effects with a negative differential effect
passed to \code{call_significant_response} (default "Negative").}

\item{pos_type}{Label for significant effects with a positive differential effect
passed to \code{call_significant_response} (default "Positive").}
}
\description{
Plots replicate comparisons for all replicates in a list of screens and outputs
plots to a given folder. Works for data returned from \code{call_significant_response}.
}
