% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_lfc_combn}
\alias{plot_lfc_combn}
\title{Plot LFCs for all gene pairs.}
\usage{
plot_lfc_combn(
  scores,
  residuals,
  condition_name,
  output_folder,
  neg_type = "Negative",
  pos_type = "Positive"
)
}
\arguments{
\item{scores}{Dataframe of scores returned from \code{call_significant_response_combn}.}

\item{residuals}{Residuals returned with the return_residuals argument set to true
from \code{call_significant_response_combn}.}

\item{condition_name}{Name of condition passed to \code{call_significant_response_combn}.}

\item{output_folder}{Folder to output plots to.}

\item{neg_type}{Label for significant effects with a negative differential effect
passed to \code{call_significant_response_combn} (default "Negative").}

\item{pos_type}{Label for significant effects with a positive differential effect
passed to \code{call_significant_response_combn} (default "Positive").}
}
\description{
Plots replicate comparisons for all replicates in a list of screens and outputs
plots to a given folder. Works for data returned from \code{call_significant_response_combn}.
}
