% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_lfc_qc}
\alias{plot_lfc_qc}
\title{Plot replicate comparisons.}
\usage{
plot_lfc_qc(df, screens, output_folder, plot_type = "png")
}
\arguments{
\item{df}{Reads or lfc dataframe.}

\item{screens}{List of screens created with \code{add_screens}.}

\item{output_folder}{Folder to output plots to.}

\item{plot_type}{Type of plot to output, one of "png" or "pdf" (default "png").}
}
\description{
Plots replicate comparisons for all replicates in a list of screens and outputs
plots to a given folder.
}
