% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_reads}
\alias{plot_reads}
\title{Plot read counts.}
\usage{
plot_reads(df, col, log_scale = TRUE, pseudocount = 1)
}
\arguments{
\item{df}{Reads dataframe.}

\item{col}{Name of column to plot.}

\item{log_scale}{If true, log-normalizes data.}

\item{pseudocount}{Pseudocounts to add to log-normalized data if specified (default 1).}
}
\value{
A ggplot object.
}
\description{
Plots a histogram of read counts for a given column.
}
