% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_rep_comparisons}
\alias{plot_rep_comparisons}
\title{Plot replicate comparisons.}
\usage{
plot_rep_comparisons(df, screens, output_folder)
}
\arguments{
\item{df}{Reads or lfc dataframe.}

\item{screens}{List of screens created with \code{add_screens}.}

\item{output_folder}{Folder to output plots to.}
}
\description{
Plots replicate comparisons for all replicates in a list of screens and outputs
plots to a given folder.
}
