% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_samples}
\alias{plot_samples}
\title{Plots sample comparisons.}
\usage{
plot_samples(
  df,
  xcol,
  ycol,
  xlab,
  ylab,
  color_col = NULL,
  color_lab = NULL,
  print_cor = FALSE
)
}
\arguments{
\item{df}{Reads or lfc dataframe.}

\item{xcol}{Name of column containing values to plot on the x-axis.}

\item{ycol}{Name of column containing values to plot on the y-axis.}

\item{color_col}{Name of column to color points by (optional).}

\item{color_lab}{Name of color legend (optional, defaults to color_col).}

\item{print_cor}{If true, prints Pearson correlation between columns 
(default FALSE).}
}
\value{
A ggplot object.
}
\description{
Pretty-plots comparisons between two samples in a scatterplot.
}
