% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_screen_reads}
\alias{plot_screen_reads}
\title{Plot read counts for a screen.}
\usage{
plot_screen_reads(
  df,
  screens,
  output_folder,
  log_scale = TRUE,
  pseudocount = 1
)
}
\arguments{
\item{df}{Reads dataframes.}

\item{screens}{List of screens created with \code{add_screens}.}

\item{output_folder}{Folder to output plots to.}

\item{log_scale}{If true, log-normalizes data.}

\item{pseudocount}{Pseudocounts to add to log-normalized data if specified (default 1).}
}
\description{
Plots a histogram of read counts for each replicate of all screens.
}
