% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_significant_response}
\alias{plot_significant_response}
\title{Plots drug response for scored data.}
\usage{
plot_significant_response(scores, control_name, condition_name, loess = TRUE)
}
\arguments{
\item{scores}{Dataframe of scores returned from \code{call_significant_response}.}

\item{control_name}{Name of control passed to \code{call_significant_response}.}

\item{condition_name}{Name of condition passed to \code{call_significant_response}.}

\item{loess}{If true and data was loess-normalized, plots loess null model instead
(default TRUE).}
}
\value{
A ggplot object.
}
\description{
Pretty-plots response for data which does not use a derived null model (e.g. for directly
comparing drug response to DMSO response). Assumes that data was scored by 
\code{score_conditions_vs_control} and significant effects were called by 
\code{call_significant_response}.
}
