% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_significant_response_dual}
\alias{plot_significant_response_dual}
\title{Plots drug response for scored paired data.}
\usage{
plot_significant_response_dual(scores, condition_name, filter_name = NULL)
}
\arguments{
\item{scores}{Dataframe of scores returned from \code{call_significant_response_dual}.}

\item{condition_name}{Name of condition passed to \code{call_significant_response_dual}.}

\item{filter_name}{If specified, calls points as non-significant if they are significant
in this column (e.g. to remove points significant in DMSO screens). Default NULL.}
}
\value{
A ggplot object.
}
\description{
Pretty-plots response for data which uses a derived null model (e.g. for comparing
dual-gene knockout effects to a multiplicative null model derived from single-gene
effects). Assumes that data was scored by \code{score_dual_vs_single} and 
significant effects were called by \code{call_significant_response_dual}.
}
