% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_screen}
\alias{remove_screen}
\title{Removes a screen from a list of screens}
\usage{
remove_screen(screen_list, name)
}
\arguments{
\item{screen_list}{An existing list of screens.}

\item{name}{A name for the screen to remove (e.g. "RPE1_T18").}
}
\value{
\code{screen_list} with the named screen removed.
}
\description{
Removes a screen with a given name from a list of screens. If multiple 
screens match the provided name, they are all removed.
}
