% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_combn_batch}
\alias{score_combn_batch}
\title{Scores multiple combinatorial screens against derived null models}
\usage{
score_combn_batch(
  combn_guides,
  single_guides,
  screens,
  batch_table,
  output_folder,
  separate_orientation = FALSE,
  min_guides = 3,
  test = "moderated-t",
  loess = TRUE,
  filter_genes = NULL,
  ignore_orientation = FALSE,
  fdr_method = "BY",
  fdr_threshold = 0.1,
  differential_threshold = 0.5,
  neg_type = "Negative",
  pos_type = "Positive",
  plot_type = "png"
)
}
\arguments{
\item{combn_guides}{A list of exonic-exonic guides returned from \code{split_guides}.}

\item{single_guides}{A list of exonic-intergenic guides returned from \code{split_guides}.}

\item{screens}{List of screens generated with \code{add_screens}.}

\item{batch_table}{Either a dataframe or a path to .tsv file mapping screens to their controls 
for scoring, with two columns for "Screen" and "Control." Screens to score against derived 
null-models with the combn scoring mode must have their respective control labeled as "combn."}

\item{output_folder}{Folder to output scored data and plots to.}

\item{separate_orientation}{If true, then guide values are scored separately across each 
orientation (default FALSE).}

\item{min_guides}{The minimum number of guides per gene pair required to score data 
(default 3).}

\item{test}{Type of hypothesis testing to run. Must be one of "rank-sum" for Wilcoxon
rank-sum testing or "moderated-t" for moderated t-testing (default "moderated-t").}

\item{loess}{If true, loess-normalizes residuals before running hypothesis testing.
Only works when test = "moderated-t" (default TRUE).}

\item{filter_genes}{List of genes to filter from scoring (default NULL).}

\item{ignore_orientation}{If TRUE, aggregates guides across both orientations, returning only
one p-value and FDR column with orientation2 p-values set to NA (default FALSE).}

\item{fdr_method}{Type of FDR to compute. One of "BH", "BY" or "bonferroni" (default
"BY")}

\item{fdr_threshold}{Threshold below which to call gene effects as significant 
(default 0.1).}

\item{differential_threshold}{Absolute value threshold on differential effects, 
below which gene effects are not called as significant (default 0.5).}

\item{neg_type}{Label for significant effects with a negative differential effect
(default "Negative").}

\item{pos_type}{Label for significant effects with a positive differential effect
(default "Positive").}

\item{plot_type}{Type of plot to output, one of "png" or "pdf" (default "png").}
}
\description{
Takes in an input .tsv file with two columns for "Screen" and "Control" and scores
all screens listed in "Screen" against their corresponding screens listed in
"Control." Outputs all files and plots in the specified folder.
}
