% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_conditions_batch}
\alias{score_conditions_batch}
\title{Scores multiple conditions against multiple controls}
\usage{
score_conditions_batch(
  guides,
  screens,
  batch_file,
  output_folder,
  separate_orientation = FALSE,
  min_guides = 3,
  test = "moderated-t",
  loess = TRUE,
  filter_genes = NULL,
  fdr_method = "BY",
  fdr_threshold = 0.1,
  differential_threshold = 0.5,
  neg_type = "Negative",
  pos_type = "Positive",
  plot_type = "png"
)
}
\arguments{
\item{guides}{A list of guides returned from \code{retrieve_guides_by_label} or 
from an entry of \code{split_guides_by_type}.}

\item{screens}{List of screens generated with \code{add_screens}.}

\item{batch_file}{Path to .tsv file mapping screens to their controls for scoring, with two 
columns for "Screen" and "Control." Screens to score against dervied null-models with the
combn scoring mode must have their respective control labeled as "combn."}

\item{output_folder}{Folder to output scored data and plots to.}

\item{separate_orientation}{If true, then guide values are scored separately across each 
orientation (default FALSE).}

\item{min_guides}{The minimum number of guides per gene pair required to score data 
(default 3).}

\item{test}{Type of hypothesis testing to run. Must be one of "rank-sum" for Wilcoxon
rank-sum testing or "moderated-t" for moderated t-testing (default "moderated-t").}

\item{loess}{If true, loess-normalizes residuals before running hypothesis testing.
Only works when test = "moderated-t" (default TRUE).}

\item{filter_genes}{List of genes to filter from scoring (default NULL).}

\item{fdr_method}{Type of FDR to compute. One of "BH", "BY" or "bonferroni" (default
"BY")}

\item{fdr_threshold}{Threshold below which to call gene effects as significant 
(default 0.1).}

\item{differential_threshold}{Absolute value threshold on differential effects, 
below which gene effects are not called as significant (default 0.5).}

\item{neg_type}{Label for significant effects with a negative differential effect
(default "Negative").}

\item{pos_type}{Label for significant effects with a positive differential effect
(default "Positive").}

\item{plot_type}{Type of plot to output, one of "png" or "pdf" (default "png").}
}
\description{
Takes in an input .tsv file with two columns for "Screen" and "Control" and scores
all screens listed in "Screen" against their corresponding screens listed in
"Control." Outputs all files and plots in the specified folder.
}
