% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_conditions_vs_control}
\alias{score_conditions_vs_control}
\title{Scores conditions against a single control.}
\usage{
score_conditions_vs_control(guides, control_cols, ..., min_guides = 3)
}
\arguments{
\item{guides}{A list of guides returned from \code{retrieve_guides_by_label} or 
from an entry of \code{split_guides_by_type}.}

\item{control_cols}{A list where the first entry is a desired name for a condition, 
e.g. DMSO, and all subsequent values are column names of the technical replicates 
for that condition. All technical replicates specified must have been passed to
\code{retrieve_guides_by_gene} or \code{retrieve_guides_by_gene} in the guide_cols
argument. For example, c("DMSO", "DMSO_T15A", "DMSO_T15B", "DMSO_T15C").}

\item{...}{Any number of conditions to score against the specified control, passed 
in as separate lists formatted the same as control_cols.}

\item{min_guides}{The minimum number of guides per gene pair required to score data 
(default 3).}
}
\value{
A dataframe of scored data with separate columns given by the specified control
  and condition names.
}
\description{
Scores guides for any number of condition screens against a control screen
(e.g. for directly comparing drug response to DMSO response). Do NOT use 
this function if you would like to compare guides against a null model 
derived from single-gene effects. In that case, call \code{score_dual_vs_single}
instead. After running this function, pass the resulting dataframe to 
\code{call_significant_response} to call significant effects.
}
