% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_dual_vs_single}
\alias{score_dual_vs_single}
\title{Scores conditions against a single control.}
\usage{
score_dual_vs_single(
  dual_guides,
  single_guides,
  ...,
  test = "rank-sum",
  min_guides = 3
)
}
\arguments{
\item{dual_guides}{A list of exonic-exonic guides returned from \code{retrieve_guides_by_label} 
or from the exonic-exonic entry of \code{split_guides_by_type}.}

\item{single_guides}{A list of exonic-intergenic guides returned from \code{retrieve_guides_by_label} 
or from the exonic-intergenic entry of \code{split_guides_by_type}.}

\item{...}{Any number of lists to score against a multiplicative null model derived 
from single-gene effects. Each list must be formatted where the first entry is a 
desired name for a condition, e.g. DMSO, and all subsequent values are column names 
of the technical replicates for that condition. All technical replicates specified 
must have been passed to \code{retrieve_guides_by_gene} or 
\code{retrieve_guides_by_label} in the guide_cols argument. For example, 
\code{c("DMSO", "DMSO_T15A", "DMSO_T15B", "DMSO_T15C")}.}

\item{test}{The desired type of hypothesis testing to run. Must be one of "rank-sum" or
"moderated-t".}

\item{min_guides}{The minimum number of guides per gene pair required to score data 
(default 3).}
}
\value{
A dataframe of scored data with separate columns given by the specified control
  and condition names.
}
\description{
Scores guides for any number of condition screens against a multiplicative null 
model derived from single-gene effects (e.g. for comparing paralog double-knockouts
to single-knockouts for each paralogous gene). Do NOT use this function if you 
would like to directly compare conditions against a control. In that case, call 
\code{score_conditions_vs_control} instead. After running this function, pass the 
resulting dataframe to \code{call_significant_response_dual} to call significant 
effects.
}
