% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{split_guides}
\alias{split_guides}
\title{Splits data into different types of guides}
\usage{
split_guides(guides, screens, id_col1 = NULL, id_col2 = NULL)
}
\arguments{
\item{guides}{Dataframe returned from \code{retrieve_guides_by_gene}.}

\item{screens}{List of screens generated with \code{add_screens}.}

\item{id_col1}{A column containing guide-specific indices, e.g. guide sequences or IDs, for
guides in the gene1 column (optional, default NULL).}

\item{id_col2}{A column containing guide-specific indices, e.g. guide sequences or IDs, for
guides in the gene1 column (optional, default NULL).}
}
\value{
A list of three separate guide lists. Dual-gene exonic-exonic guides are
  stored in the key "exonic_exonic", exonic-intergenic guides are stored in the 
  key "exonic_intergenic" and single-gene exonic-exonic guides are stored in the
  key "single_gene_dual_targeted".
}
\description{
Splits data into different types of guides
}
