% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_screen}
\alias{add_screen}
\title{Adds a new screen}
\usage{
add_screen(screen_list = NULL, name, replicates, normalize_name = NULL)
}
\arguments{
\item{screen_list}{An existing list of screens (optional, default NULL).}

\item{name}{A name for the screen (e.g. "RPE1_T18").}

\item{replicates}{A list of columns containing replicate data for the given screen.}

\item{normalize_name}{A name of another screen to normalize data in this screen to
(e.g. "RPE1_T0").}
}
\value{
A named list corresponding to provided screen names, where each sub-list 
  contains a list of the replicate columns (in "replicates") and the screen to 
  normalize against (in "normalize_name").
}
\description{
Makes a list containing info for a given screen and optionally appends it to a given list.
}
