% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_combn_vs_single}
\alias{score_combn_vs_single}
\title{Scores conditions against a single control.}
\usage{
score_combn_vs_single(
  combn_guides,
  single_guides,
  screens,
  screen_names,
  min_guides = 3,
  test = "moderated-t",
  loess = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{combn_guides}{A list of exonic-exonic guides returned from \code{retrieve_guides_by_label} 
or from the exonic-exonic entry of \code{split_guides_by_type}.}

\item{single_guides}{A list of exonic-intergenic guides returned from \code{retrieve_guides_by_label} 
or from the exonic-intergenic entry of \code{split_guides_by_type}.}

\item{screens}{List of screens generated with \code{add_screens}.}

\item{screen_names}{A list of screen names to score against a derived null model from
single-gene effects.}

\item{min_guides}{The minimum number of guides per gene pair required to score data 
(default 3).}

\item{test}{Type of hypothesis testing to run. Must be one of "rank-sum" for Wilcoxon
rank-sum testing or "moderated-t" for moderated t-testing (default "moderated-t").}

\item{loess}{If true, loess-normalizes residuals before running hypothesis testing.
Only works when test = "moderated-t" (default TRUE).}

\item{verbose}{If true, prints verbose output (default FALSE).}
}
\value{
A dataframe of scored data with separate columns given by the specified control
  and condition names.
}
\description{
Scores guides for any number of condition screens against a multiplicative null 
model derived from single-gene effects (e.g. for comparing paralog double-knockouts
to single-knockouts for each paralogous gene). Do NOT use this function if you 
would like to directly compare conditions against a control. In that case, call 
\code{score_conditions_vs_control} instead. After running this function, pass the 
resulting dataframe to \code{call_significant_response_dual} to call significant 
effects.
}
