% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_conditions_vs_control}
\alias{score_conditions_vs_control}
\title{Scores conditions against a single control.}
\usage{
score_conditions_vs_control(
  guides,
  screens,
  control_screen_name,
  condition_screen_names,
  orientation = FALSE,
  min_guides = 3,
  test = "moderated-t",
  loess = TRUE,
  return_residuals = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{guides}{A list of guides returned from \code{retrieve_guides_by_label} or 
from an entry of \code{split_guides_by_type}.}

\item{screens}{List of screens generated with \code{add_screens}.}

\item{control_screen_name}{Name of a control screen to test condition screens against.}

\item{condition_screen_names}{A list of condition screen names to score against the 
control screen.}

\item{orientation}{If true, then guide values are scored separately across each 
orientation (default FALSE).}

\item{min_guides}{The minimum number of guides per gene pair required to score data 
(default 3).}

\item{test}{Type of hypothesis testing to run. Must be one of "rank-sum" for Wilcoxon
rank-sum testing or "moderated-t" for moderated t-testing (default "moderated-t").}

\item{loess}{If true, loess-normalizes residuals before running hypothesis testing.
Only works when test = "moderated-t" (default TRUE).}

\item{return_residuals}{If true, additionally returns residuals dataframe (default FALSE).}

\item{verbose}{If true, prints verbose output (default FALSE).}
}
\value{
A dataframe of scored data with separate columns given by the specified control
  and condition names. If return_residuals is true, instead returns a list where the
  first element is the scored data and the second is the residuals dataframe.
}
\description{
Scores guides for any number of condition screens against a control screen
(e.g. for directly comparing drug response to DMSO response). Do NOT use 
this function if you would like to compare guides against a null model 
derived from single-gene effects. In that case, call \code{score_combn_vs_single}
instead. After running this function, pass the resulting dataframe to 
\code{call_significant_response} to call significant effects.
}
